Linter

+++ apps/blender/task/blendercropper.py
@@ -59,22 +59,27 @@ class BlenderCropper:
         useful for cropping with blender, second one are corresponding
         pixels. Each list has splits_num elements, one for each split.
         """
-        logger.info("Values left=%r, right=%r, top=%r, bottom=%r", image_border[0],
-                    image_border[1], image_border[3], image_border[2])
+        logger.info("Values left=%r, right=%r, top=%r, bottom=%r",
+                    image_border[0], image_border[1], image_border[3],
+                    image_border[2])
 
         #  This is how Blender is calculating pixel check
         #  BlenderSync::get_buffer_params in blender_camers.cpp file
         #  BoundBox2D border = cam->border.clamp();
-		#  params.full_x = (int)(border.left * (float)width);
+        #  params.full_x = (int)(border.left * (float)width);
         #
         #  NOTE BLENDER IS USING FLOATS FALUES
         #  that means single precision 4 bytes floats, python is not
         #  it is using double precision values. Here numpy is used to emulate
         #  that loss of precision when assigning double to float.
-        left_p = math.floor(numpy.float32(image_border[0]) * numpy.float32(resolution[0]))
-        right_p = math.floor(numpy.float32(image_border[1]) * numpy.float32(resolution[0]))
-        bottom_p = math.floor(numpy.float32(image_border[2]) * numpy.float32(resolution[1]))
-        top_p = math.floor(numpy.float32(image_border[3]) * numpy.float32(resolution[1]))
+        left_p = math.floor(numpy.float32(image_border[0]) *
+                            numpy.float32(resolution[0]))
+        right_p = math.floor(numpy.float32(image_border[1]) *
+                             numpy.float32(resolution[0]))
+        bottom_p = math.floor(numpy.float32(image_border[2]) *
+                              numpy.float32(resolution[1]))
+        top_p = math.floor(numpy.float32(image_border[3]) *
+                           numpy.float32(resolution[1]))
 
         logger.info("Pixels left=%r, right=%r, top=%r, bottom=%r", left_p,
                     right_p, top_p, bottom_p)
@@ -97,10 +102,14 @@ class BlenderCropper:
             split_y = self._random_split(bottom_p, top_p, crop_size[1])
 
             # Here another conversion from double to float
-            x_f = numpy.float32(split_x[0] / resolution[0])
-            y_f = numpy.float32(split_x[1] / resolution[0])
-            right_f = numpy.float32(split_y[0] / resolution[1])
-            bottom_f = numpy.float32(split_y[1] / resolution[1])
+            x_f = numpy.float32(numpy.float32(split_x[0])
+                                / numpy.float32(resolution[0]))
+            y_f = numpy.float32(numpy.float32(split_x[1]) /
+                                numpy.float32(resolution[0]))
+            right_f = numpy.float32(numpy.float32(split_y[0])
+                                    / numpy.float32(resolution[1]))
+            bottom_f = numpy.float32(numpy.float32(split_y[1])
+                                     / numpy.float32(resolution[1]))
 
             # Recalculate pixel after converting to float
             split_x[0] = math.floor(x_f * numpy.float32(resolution[0]))
